Travaux sur l'asservissement une semaine avant la coupe

+++ Includes/RessourcesHardware.h
@@ -20,8 +20,8 @@ void CdeMoteur2(float cde_pourcent);
 void CdeServo1(int ppm1500);
 void CdeServo2(int ppm1500);
 void CdeServo3(int ppm1500);
-int getCodeur1();
-int getCodeur2();
+short getCodeur1();
+short getCodeur2();
 unsigned int readAnalog(int channel);
 float readAnalogVolt(int channel);
